third_party/WebKit/LayoutTests/hittesting/image-with-border-radius-expected.txt

The border radius clip should be respected when hit testing images.


PASS successfullyParsed is true

TEST COMPLETE
PASS document.elementFromPoint(x + 20, y + 20).id is 'container'
PASS document.elementFromPoint(x + 35, y + 35).id is 'roundedImage'
PASS document.elementFromPoint(x + 60, y + 60).id is 'roundedImage'
PASS document.elementFromPoint(x + 68, y + 68).id is 'roundedImage'
PASS document.elementFromPoint(x + 80, y + 80).id is 'roundedImage'
PASS document.elementFromPoint(x + 230, y + 230).id is 'roundedImage'
PASS document.elementFromPoint(x + 240, y + 240).id is 'roundedImage'
PASS document.elementFromPoint(x + 265, y + 265).id is 'roundedImage'
PASS document.elementFromPoint(x + 275, y + 275).id is 'container'
